Favorites: confirm on delete; swipe-left to delete; loader

+++ www/js/controllers.js
@@ -126,31 +126,55 @@ angular.module('conFusion.controllers', [])
 
 
 
-    .controller('FavoritesController', ['$scope', 'menuFactory', 'favoriteFactory', 'baseURL', '$ionicListDelegate', 
-        function ($scope, menuFactory, favoriteFactory, baseURL, $ionicListDelegate) {
+    .controller('FavoritesController', ['$scope', 'menuFactory', 'favoriteFactory', 'baseURL', '$ionicListDelegate', '$ionicPopup', '$ionicLoading', '$timeout', 
+        function ($scope, menuFactory, favoriteFactory, baseURL, $ionicListDelegate, $ionicPopup, $ionicLoading, $timeout) {
 
-            $scope.baseURL = baseURL;      
-            $scope.shouldShowDelete = false;      
-            
-            $scope.favorites = favoriteFactory.getFavorites();
-            
-            menuFactory.getDishes().query(
-                function(response) {
-                    $scope.dishes = response;
-                    console.log("Dishes: " + $scope.dishes);
-                },
-                function(response) {
-                    $scope.message = "Error: "+response.status + " " + response.statusText;
+        $scope.baseURL = baseURL;      
+        $scope.shouldShowDelete = false;      
+
+        $ionicLoading.show({
+            template: '<ion-spinner></ion-spinner> Loading...'
+        });
+        
+        $scope.favorites = favoriteFactory.getFavorites();
+        
+        menuFactory.getDishes().query(
+            function(response) {
+                $scope.dishes = response;
+                $timeout(function () {
+                    $ionicLoading.hide();
+                }, 1000);
+            },
+            function(response) {
+                $scope.message = "Error: "+response.status + " " + response.statusText;
+                $timeout(function () {
+                    $ionicLoading.hide();
+                }, 1000);
+        });
+
+        $scope.toggleDelete = function() {
+            $scope.shouldShowDelete = !$scope.shouldShowDelete;
+        }
+
+
+       $scope.deleteFavorite = function (index) { 
+
+            var confirmPopup = $ionicPopup.confirm({
+                title: 'Confirm Delete',
+                template: 'Are you sure you want to delete this item?'
             });
 
-            $scope.toggleDelete = function() {
-                $scope.shouldShowDelete = !$scope.shouldShowDelete;
-            }
+            confirmPopup.then(function (res) {
+                if (res) {
+                    console.log('Ok to delete');
+                    favoriteFactory.deleteFromFavorites(index);
+                } else {
+                    console.log('Canceled delete');
+                }
+            });
 
-           $scope.deleteFavorite = function (index) {        
-                favoriteFactory.deleteFromFavorites(index);
-                $scope.shouldShowDelete = false;
-            }
+            $scope.shouldShowDelete = false;
+        }
 
     }])
